store films saved to my list in a database for further storage
the films saved by each profile will now be added into the database under the profile ID. the new code alters the database to allow for this. for some reason though, i cant seem to get the saved movies tho show up in the list once logging back in again, despite the fact they are clearly in the database. will need further work and review